✨ implement sign in and register as popup
